+++ R/HybridRecommender.R
@@ -2,15 +2,17 @@
 
 
 HybridRecommender <- function(..., weights = NULL) {
-  recommender <- list(...)
+  recommenders <- list(...)
 
-  if(is.null(weights)) weights <- rep(1, length(recommender))
-  else if(length(recommender) != length(weights)) stop("Number of recommender and length of weights does not agree!")
+  if(length(recommenders) < 1) stop("No base recommender specified!")
+
+  if(is.null(weights)) weights <- rep(1, length(recommenders))
+  else if(length(recommenders) != length(weights)) stop("Number of recommenders and length of weights do not agree!")
   weights <- weights/sum(weights)
 
-  if(!all(sapply(recommender, is, "Recommender"))) stop("Not all supplied models are of class 'Recommender'.")
+  if(!all(sapply(recommenders, is, "Recommender"))) stop("Not all supplied models are of class 'Recommender'.")
 
-  model <- list(recommender = recommender, weights = weights)
+  model <- list(recommenders = recommenders, weights = weights)
 
   predict <- function(model=NULL, newdata, n=10,
     data= NULL, type=c("topNList", "ratings", "ratingMatrix"), ...) {
@@ -26,7 +28,7 @@ HybridRecommender <- function(..., weights = NULL) {
 
     #if(ncol(newdata) != length(model$labels)) stop("number of items in newdata does not match model.")
 
-    pred <- lapply(model$recommender, FUN = function(object)
+    pred <- lapply(model$recommenders, FUN = function(object)
       object@predict(object@model, newdata, data=data, type="ratings", ...))
 
     ratings <- matrix(NA, nrow=nrow(newdata), ncol = ncol(newdata))
@@ -43,10 +45,8 @@ HybridRecommender <- function(..., weights = NULL) {
 
     dimnames(ratings) <- dimnames(newdata)
 
-
     ratings <- as(ratings, "realRatingMatrix")
     colnames(ratings) <- colnames(newdata)
-
 
     if(type == "ratingMatrix")
       stop("Hybrid cannot predict a complete ratingMatrix!")
@@ -57,9 +57,35 @@ HybridRecommender <- function(..., weights = NULL) {
   ## this recommender has no model
   new("Recommender", method = "HYBRID",
     dataType = "ratingMatrix",
-    ntrain = NA_integer_,
+    ntrain = recommenders[[1]]@ntrain,  ### take training set size from firs recommender
     model = model,
     predict = predict)
 }
 
 
+## recommender interface
+.HYBRID_params <- list(
+  recommenders = NULL,
+  weights = NULL
+)
+
+HYBRID <- function(data, parameter = NULL) {
+  p <- getParameters(.HYBRID_params, parameter)
+
+  # build the individual recommenders
+  recommenders <- lapply(parameter$recommenders, FUN = function(p)
+    Recommender(data = data, method = p$name, parameter = p$param))
+
+  do.call(HybridRecommender, c(recommenders, weights = list(p$weights)))
+}
+
+## register recommender
+recommenderRegistry$set_entry(
+  method="HYBRID", dataType = "realRatingMatrix", fun=HYBRID,
+  description="Hybrid recommender that aggegates several recommendation strategies using weighted averages.",
+  parameters=.HYBRID_params)
+
+recommenderRegistry$set_entry(
+  method="HYBRID", dataType = "binaryRatingMatrix", fun=HYBRID,
+  description="Hybrid recommender that aggegates several recommendation strategies using weighted averages.",
+  parameters=.HYBRID_params)